refactor(services): renamed core_services to use_cases
