+++ app/api/fast_api/routes/endpoints/http/auth_controller.py
@@ -5,7 +5,7 @@ from fastapi.security import OAuth2PasswordRequestForm
 
 from app.dependencies.services.user import get_user_service
 from app.infrastructure.database.models.user import User
-from app.services.core_services.user_service import UserService
+from app.services.use_cases.user_service import UserService
 from app.dependencies.auth import get_current_active_user
 from app.infrastructure.settings.security import create_access_token, verify_password
 from app.infrastructure.settings.config import ACCESS_TOKEN_EXPIRE_DAYS
