+++ app/dependencies/auth.py
@@ -6,7 +6,7 @@ from fastapi.security import OAuth2PasswordBearer
 
 from app.infrastructure.database.models.user import User
 from app.dependencies.services.user import get_user_service
-from app.services.core_services.user_service import UserService
+from app.services.use_cases.user_service import UserService
 from app.infrastructure.settings.config import AUTH_KEY, HASHING_ALGORITHM
