+++ app/dependencies/services/comment.py
@@ -1,6 +1,6 @@
 from fastapi import Depends
 
-from app.services.core_services.comment_service import CommentService
+from app.services.use_cases.comment_service import CommentService
 from app.infrastructure.database.repositories.comment_repository import CommentRepository
 from app.dependencies.db import get_db
