+++ app/dependencies/services/subscription.py
@@ -1,6 +1,6 @@
 from fastapi import Depends
 
-from app.services.core_services.subscription_service import SubscriptionService
+from app.services.use_cases.subscription_service import SubscriptionService
 from app.infrastructure.database.repositories.subscription_repository import SubscriptionRepository
 from app.dependencies.db import get_db
