+++ app/infrastructure/settings/config.py
@@ -27,3 +27,6 @@ ALLOWED_HOSTS = [
     'http://127.0.0.1:8000',
     'http://127.0.0.1:5500'
 ]
+
+# REDIS_URL = os.getenv('REDIS_URL')
+REDIS_URL = "redis://127.0.0.1:6379"
